+++ handlers/whatsapp/whatsapp.go
@@ -23,6 +23,10 @@ const (
 	configNamespace = "fb_namespace"
 )
 
+var (
+	retryParam = ""
+)
+
 func init() {
 	courier.RegisterHandler(newHandler())
 }
@@ -166,19 +170,19 @@ func (h *handler) receiveEvent(ctx context.Context, channel courier.Channel, w h
 
 		if msg.Type == "text" {
 			text = msg.Text.Body
-		} else if msg.Type == "audio" {
+		} else if msg.Type == "audio" && msg.Audio != nil {
 			mediaURL, err = resolveMediaURL(channel, msg.Audio.ID)
-		} else if msg.Type == "document" {
+		} else if msg.Type == "document" && msg.Document != nil {
 			text = msg.Document.Caption
 			mediaURL, err = resolveMediaURL(channel, msg.Document.ID)
-		} else if msg.Type == "image" {
+		} else if msg.Type == "image" && msg.Image != nil {
 			text = msg.Image.Caption
 			mediaURL, err = resolveMediaURL(channel, msg.Image.ID)
-		} else if msg.Type == "location" {
+		} else if msg.Type == "location" && msg.Location != nil {
 			mediaURL = fmt.Sprintf("geo:%f,%f", msg.Location.Latitude, msg.Location.Longitude)
-		} else if msg.Type == "video" {
+		} else if msg.Type == "video" && msg.Video != nil {
 			mediaURL, err = resolveMediaURL(channel, msg.Video.ID)
-		} else if msg.Type == "voice" {
+		} else if msg.Type == "voice" && msg.Voice != nil {
 			mediaURL, err = resolveMediaURL(channel, msg.Voice.ID)
 		} else {
 			// we received a message type we do not support.
@@ -360,6 +364,14 @@ type mtVideoPayload struct {
 	Video *mediaObject `json:"video"`
 }
 
+type mtErrorPayload struct {
+	Errors []struct {
+		Code    int    `json:"code"`
+		Title   string `json:"title"`
+		Details string `json:"details"`
+	} `json:"errors"`
+}
+
 // whatsapp only allows messages up to 4096 chars
 const maxMsgLength = 4096
 
@@ -378,10 +390,9 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 		return nil, fmt.Errorf("invalid base url set for WA channel: %s", err)
 	}
 	sendPath, _ := url.Parse("/v1/messages")
-	sendURL := url.ResolveReference(sendPath).String()
 
 	status := h.Backend().NewMsgStatusForID(msg.Channel(), msg.ID(), courier.MsgErrored)
-	var log *courier.ChannelLog
+	var logs []*courier.ChannelLog
 
 	if len(msg.Attachments()) > 0 {
 		for attachmentCount, attachment := range msg.Attachments() {
@@ -395,7 +406,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 					Type: "audio",
 				}
 				payload.Audio = &mediaObject{Link: s3url}
-				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
+				externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
 
 			} else if strings.HasPrefix(mimeType, "application") {
 				payload := mtDocumentPayload{
@@ -408,7 +419,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				} else {
 					payload.Document = &mediaObject{Link: s3url}
 				}
-				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
+				externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
 
 			} else if strings.HasPrefix(mimeType, "image") {
 				payload := mtImagePayload{
@@ -420,7 +431,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				} else {
 					payload.Image = &mediaObject{Link: s3url}
 				}
-				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
+				externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
 			} else if strings.HasPrefix(mimeType, "video") {
 				payload := mtVideoPayload{
 					To:   msg.URN().Path(),
@@ -431,15 +442,15 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				} else {
 					payload.Video = &mediaObject{Link: s3url}
 				}
-				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
+				externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
 			} else {
 				duration := time.Since(start)
 				err = fmt.Errorf("unknown attachment mime type: %s", mimeType)
-				log = courier.NewChannelLogFromError("Error sending message", msg.Channel(), msg.ID(), duration, err)
+				logs = []*courier.ChannelLog{courier.NewChannelLogFromError("Error sending message", msg.Channel(), msg.ID(), duration, err)}
 			}
 
-			// if we have a log, add it to our status
-			if log != nil {
+			// add logs to our status
+			for _, log := range logs {
 				status.AddLog(log)
 			}
 
@@ -481,12 +492,14 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 			}
 
 			externalID := ""
-			externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
-			status.AddLog(log)
+			externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
 
-			if err != nil {
-				log.WithError("Error sending message", err)
-			} else {
+			// add logs to our status
+			for _, log := range logs {
+				status.AddLog(log)
+			}
+
+			if err == nil {
 				status.SetExternalID(externalID)
 			}
 		} else {
@@ -498,12 +511,13 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 					Type: "text",
 				}
 				payload.Text.Body = part
+				externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
 
-				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
-				status.AddLog(log)
-
+				// add logs to our status
+				for _, log := range logs {
+					status.AddLog(log)
+				}
 				if err != nil {
-					log.WithError("Error sending message", err)
 					break
 				}
 
@@ -523,38 +537,124 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 	return status, nil
 }
 
-func sendWhatsAppMsg(msg courier.Msg, url string, token string, payload interface{}) (string, *courier.ChannelLog, error) {
+func sendWhatsAppMsg(msg courier.Msg, sendPath *url.URL, token string, payload interface{}) (string, []*courier.ChannelLog, error) {
+	start := time.Now()
 	jsonBody, err := json.Marshal(payload)
-	if err != nil {
-		log := courier.NewChannelLog("unable to build JSON body", msg.Channel(), msg.ID(), "", "", courier.NilStatusCode, "", "", time.Duration(0), err)
-		return "", log, err
-	}
 
-	req, _ := http.NewRequest(http.MethodPost, url, bytes.NewReader(jsonBody))
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Accept", "application/json")
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", token))
-	req.Header.Set("User-Agent", utils.HTTPUserAgent)
+	if err != nil {
+		elapsed := time.Now().Sub(start)
+		log := courier.NewChannelLogFromError("unable to build JSON body", msg.Channel(), msg.ID(), elapsed, err)
+		return "", []*courier.ChannelLog{log}, err
+	}
+	req, _ := http.NewRequest(http.MethodPost, sendPath.String(), bytes.NewReader(jsonBody))
+	req.Header = buildWhatsAppRequestHeader(token)
 	rr, err := utils.MakeHTTPRequest(req)
 	log := courier.NewChannelLogFromRR("Message Sent", msg.Channel(), msg.ID(), rr).WithError("Message Send Error", err)
+	errPayload := &mtErrorPayload{}
+	err = json.Unmarshal(rr.Body, errPayload)
+
+	// handle send msg errors
+	if err == nil && len(errPayload.Errors) > 0 {
+		if !hasWhatsAppContactError(*errPayload) {
+			err := errors.Errorf("received error from send endpoint: %s", errPayload.Errors[0].Title)
+			return "", []*courier.ChannelLog{log}, err
+		}
+		// check contact
+		baseURL := fmt.Sprintf("%s://%s", sendPath.Scheme, sendPath.Host)
+		rrCheck, err := checkWhatsAppContact(baseURL, token, msg.URN())
+
+		if rrCheck == nil {
+			elapsed := time.Now().Sub(start)
+			checkLog := courier.NewChannelLogFromError("unable to build contact check request", msg.Channel(), msg.ID(), elapsed, err)
+			return "", []*courier.ChannelLog{log, checkLog}, err
+		}
+		checkLog := courier.NewChannelLogFromRR("Contact check", msg.Channel(), msg.ID(), rrCheck).WithError("Status Error", err)
+
+		if err != nil {
+			return "", []*courier.ChannelLog{log, checkLog}, err
+		}
+		// try send msg again
+		reqRetry, _ := http.NewRequest(http.MethodPost, sendPath.String(), bytes.NewReader(jsonBody))
+		reqRetry.Header = buildWhatsAppRequestHeader(token)
+
+		if retryParam != "" {
+			reqRetry.URL.RawQuery = fmt.Sprintf("%s=1", retryParam)
+		}
+		rrRetry, err := utils.MakeHTTPRequest(reqRetry)
+		retryLog := courier.NewChannelLogFromRR("Message Sent", msg.Channel(), msg.ID(), rr).WithError("Message Send Error", err)
+
+		if err != nil {
+			return "", []*courier.ChannelLog{log, checkLog, retryLog}, err
+		}
+		externalID, err := getSendWhatsAppMsgId(rrRetry)
+		return externalID, []*courier.ChannelLog{log, checkLog, retryLog}, err
+	}
+	externalID, err := getSendWhatsAppMsgId(rr)
+	return externalID, []*courier.ChannelLog{log}, err
+}
+
+func buildWhatsAppRequestHeader(token string) http.Header {
+	header := http.Header{
+		"Content-Type":  []string{"application/json"},
+		"Accept":        []string{"application/json"},
+		"Authorization": []string{fmt.Sprintf("Bearer %s", token)},
+		"User-Agent":    []string{utils.HTTPUserAgent},
+	}
+	return header
+}
+
+func hasWhatsAppContactError(payload mtErrorPayload) bool {
+	for _, err := range payload.Errors {
+		if err.Code == 1006 && err.Title == "Resource not found" && err.Details == "unknown contact" {
+			return true
+		}
+	}
+	return false
+}
+
+func getSendWhatsAppMsgId(rr *utils.RequestResponse) (string, error) {
+	if externalID, err := jsonparser.GetString(rr.Body, "messages", "[0]", "id"); err == nil {
+		return externalID, nil
+	} else {
+		return "", errors.Errorf("unable to get message id from response body")
+	}
+}
+
+type mtContactCheckPayload struct {
+	Blocking   string   `json:"blocking"`
+	Contacts   []string `json:"contacts"`
+	ForceCheck bool     `json:"force_check"`
+}
+
+func checkWhatsAppContact(baseURL string, token string, urn urns.URN) (*utils.RequestResponse, error) {
+	payload := mtContactCheckPayload{
+		Blocking:   "wait",
+		Contacts:   []string{fmt.Sprintf("+%s", urn.Path())},
+		ForceCheck: true,
+	}
+	reqBody, err := json.Marshal(payload)
+
 	if err != nil {
-		return "", log, err
+		return nil, err
 	}
+	sendURL := fmt.Sprintf("%s/v1/contacts", baseURL)
+	req, _ := http.NewRequest(http.MethodPost, sendURL, bytes.NewReader(reqBody))
+	req.Header = buildWhatsAppRequestHeader(token)
+	rr, err := utils.MakeHTTPRequest(req)
 
-	errorTitle, err := jsonparser.GetString(rr.Body, "errors", "[0]", "title")
-	if errorTitle != "" {
-		err = errors.Errorf("received error from send endpoint: %s", errorTitle)
-		return "", log, err
-	}
-
-	// grab the id
-	externalID, err := jsonparser.GetString(rr.Body, "messages", "[0]", "id")
 	if err != nil {
-		err := errors.Errorf("unable to get message id from response body")
-		return "", log, err
+		return rr, err
 	}
-
-	return externalID, log, err
+	// check contact status
+	if status, err := jsonparser.GetString(rr.Body, "contacts", "[0]", "status"); err == nil {
+		if status == "valid" {
+			return rr, nil
+		} else {
+			return rr, errors.Errorf(`contact status is "%s"`, status)
+		}
+	} else {
+		return rr, err
+	}
 }
 
 func (h *handler) getTemplate(msg courier.Msg) (*MsgTemplating, error) {